refactor: uuid Interface로 분리
