+++ soridam-domain/src/main/java/sorisoop/soridam/domain/noise/domain/Noise.java
@@ -2,12 +2,13 @@ package sorisoop.soridam.domain.noise.domain;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static lombok.AccessLevel.PROTECTED;
-import static sorisoop.soridam.infra.uuid.UuidPrefix.NOISE;
+import static sorisoop.soridam.common.uuid.UuidPrefix.NOISE;
 
 import org.locationtech.jts.geom.Point;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
@@ -15,15 +16,17 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import sorisoop.soridam.common.domain.BaseTimeEntity;
+import sorisoop.soridam.common.domain.UuidExtractable;
 import sorisoop.soridam.domain.user.domain.User;
-import sorisoop.soridam.infra.uuid.PrefixedUuid;
+import sorisoop.soridam.common.uuid.PrefixedUuid;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = PROTECTED)
 @AllArgsConstructor(access = PROTECTED)
-public class Noise extends BaseTimeEntity {
+public class Noise extends BaseTimeEntity implements UuidExtractable {
+	@Id
 	@PrefixedUuid(NOISE)
 	private String id;
